refactor: adding check token apis Update auth module and resolver
Refactor the `AuthModule` and `AuthResolver` to include the new `CheckTokenInput` and `validateToken` functionality. This allows for checking the validity of a JWT token. The `AuthResolver` now includes a `checkToken` query that takes in the `CheckTokenInput` and returns a boolean indicating whether the token is valid or not.
