+++ backend/src/auth/auth.module.ts
@@ -2,10 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Menu } from './menu/menu.model';
 import { JwtModule } from '@nestjs/jwt';
-import { Role } from './dto/role.model';
+import { Role } from './role/role.model';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { AuthService } from './auth.service';
 import { User } from 'src/user/user.model';
+import { AuthResolver } from './auth.resolver';
 
 @Module({
   imports: [
@@ -20,7 +21,7 @@ import { User } from 'src/user/user.model';
       inject: [ConfigService],
     }),
   ],
-  providers: [AuthService],
+  providers: [AuthService, AuthResolver],
   exports: [AuthService],
 })
 export class AuthModule {}
